Move appropriate tests/methods into Cipher
