+++ test/enigma_test.rb
@@ -30,50 +30,6 @@ class EnigmaTest < Minitest::Test
     assert_equal expected2, @enigma.encrypt("HELLO! WORLD!", "00000", "000000")
   end
 
-  def test_it_can_get_keys
-    key = "02715"
-
-    assert_equal [2,27,71,15], @enigma.get_keys(key)
-  end
-
-  def test_key_to_pairs
-    assert_equal [10,4,45,55], @enigma.key_to_pairs("10455")
-  end
-
-  def test_date_squared_last_four_num
-    assert_equal [4,4,0,0], @enigma.date_to_keys("120620")
-  end
-
-  def test_total_offset
-    key = [1,1,1,1]
-    date = [1,1,1,1]
-    assert_equal [2,2,2,2], @enigma.total_offset(key, date)
-  end
-
-  def test_it_can_get_date_key
-    date = "040895"
-
-    assert_equal [1,0,2,5], @enigma.get_date_keys(date)
-  end
-
-  def test_it_can_get_offset_total
-    date = "040895"
-    key = "02715"
-
-    assert_equal [3,27,73,20], @enigma.get_info(key, date)
-  end
-
-  def test_shift_alphabet
-    skip
-    assert_equal 1, @enigma.shift_alphabet
-  end
-
-  def test_it_can_encrypt_a_message
-    skip
-    message = "abc"
-    assert_equal "bcd", @enigma.new_message(message)
-  end
-
   def test_it_can_decrypt_a_message
     expected =
     {
@@ -85,11 +41,6 @@ class EnigmaTest < Minitest::Test
     assert_equal expected, @enigma.decrypt("keder ohulw", "02715", "040895")
   end
 
-  def test_it_can_get_todays_date
-    expected = Date.today.strftime("%d%m%y")
-
-    assert_equal expected, @enigma.get_date
-  end
 
   def test_it_can_encrypt_with_todays_date
     Date.expects(:today).returns(Date.new(2020,02,28))
@@ -134,11 +85,4 @@ class EnigmaTest < Minitest::Test
     assert_equal expected, @enigma.encrypt("hello world")
   end
 
-  def test_it_can_generate_random_number
-    assert_instance_of String, @enigma.get_random_key
-    assert @enigma.get_random_key.to_i < 100000
-    assert_equal 5, @enigma.get_random_key.length
-    # little odd testing minimum value, as it can return "00000"
-    # assert @enigma.random_key.to_i > 999
-  end
 end
